VectorLogging.jl

The loggers in this package create log entries of this form:

LogEntry

which is conducive to automated analysis.

VectorLogger

VectorLogger creates an in-memory log which can be programaticallly analyzed.

VectorLogging.LogEntry — Type

LogEntry

LogEntry encapsulates the data for a single log message to VectorLogger.

VectorLogging.VectorLogger — Type

VectorLogger

VectorLogger is an in-memory log destination. The logger's log slot is a Vector of LogEntrys.

using Logging
using VectorLogging

logger = VectorLogger()

with_logger(logger) do
    @info "Message 1"
    @info "Message 2"
end

logger.log

2-element Vector{LogEntry}:
 LogEntry(Info, "Message 1", :Main, :index, :Main_Main_d2d482e3, "index.md", 30, Base.Pairs{Symbol, Union{}, Tuple{}, NamedTuple{(), Tuple{}}}())
 LogEntry(Info, "Message 2", :Main, :index, :Main_Main_5ac2d681, "index.md", 31, Base.Pairs{Symbol, Union{}, Tuple{}, NamedTuple{(), Tuple{}}}())

FileLogger and LogFileReader

Sometimes one needs a machine readable log that outlives a Julia session.

VectorLogging.FileLogger — Type

FileLogger

A logger that writes log entries to a file or stream in the specified format.

VectorLogging.LogFileReader — Type

LogFileReader

Provide an iteration interface to a log file written by FileLogger.

VectorLogging.AbstractLogFileFormat — Type

AbstractLogFileFormat

The supertype of all supported log file formats.

using Logging
using VectorLogging

FileLogger("foo.log", JSONLogFileFormat()) do logger
    with_logger(logger) do
        @info "Message 1"
        @info "Message 2"
    end
end

LogFileReader("foo.log", JSONLogFileFormat()) do reader
    collect(reader)
end

2-element Vector{LogEntry}:
 LogEntry(Dict{String, Any}("level" => 0), "Message 1", "Main", "index", "Main_Main_81968e01", "index.md", 56, Dict{String, Any}())
 LogEntry(Dict{String, Any}("level" => 0), "Message 2", "Main", "index", "Main_Main_d307c241", "index.md", 57, Dict{String, Any}())